+++ src/components/NorseHomepage/NorseHomepage.tsx
